move channel cotrollers under api folder
