+++ controller/index.js
@@ -1,9 +1,9 @@
 let router = require('express').Router();
 
 router.use(require('./home.js'));
-router.use(require('./channel.js'));
-router.use(require('./main_channel.js'));
-router.use(require('./main_channel_chat.js'));
+router.use(require('./api_0.0.0/channel.js'));
+router.use(require('./api_0.0.0/main_channel.js'));
+router.use(require('./api_0.0.0/main_channel_chat.js'));
 router.use(require('./api_0.0.0/user.js'));
 router.use(require('./api_0.0.0/auth.js'));
 router.use(require('./api_0.0.0/login.js'));
