+++ test/controller/api_0.0.0/manual/api_index.html
@@ -9,7 +9,7 @@
     <a href='/test/api_sign_up'>Sign up test</a><br>
     <a href='/test/api_login'>Login test</a><br>
     <a href='/test/api_logout'>logout test</a><br>
-    <a href='/channel'>Create Channel</a><br>
+    <a href='/api/0.0.0/channel'>Create Channel</a><br>
     <a href='/test/channel_connect'>Channel connection test</a><br>
     <a href='/test/main_channel_connect'>Main channel connection test</a><br>
     <a href='/test/main_channel_chat'>Main channel chat test</a><br>
