+++ test/controller/main_channel_chat.html
@@ -74,7 +74,7 @@
                 console.log('-> Sending comment');
                 return fetch
                 (
-                    '/main_channel/chat',
+                    '/api/0.0.0/main_channel/chat',
                     {
                         method : 'POST',
                         headers :
@@ -126,7 +126,7 @@
         {
             console.log('Starting a sse connection')
 
-            sse = (new EventSource('/main_channel/chat'));
+            sse = (new EventSource('/api/0.0.0/main_channel/chat'));
 
             sse.onmessage = (ev) =>
             {
